fix: move codegen spec file from outside of src to src
