+++ tsconfig.json
@@ -69,6 +69,7 @@
     "forceConsistentCasingInFileNames": true        /* Disallow inconsistently-cased references to the same file. */
   },
   "include": [
-    "./src/", "./codegenSpec"
+    "./src/",
+    "./src/codegenSpec"
   ]
 }
